requirements.txt
pyyaml
flask
flask-restful
colorama
pytest